apply feedback
1. add SampleToken Deploy Tutorial using T-Bears
2. revert is going to be duplecated soon.
3. add crowdsale's code about How to use internal call

=========== feedback ===========
Need to separate Audit checklist from this documents, since Audit lists are frequently able to be changed.
Need to explain overall process and structure of Contract. I think this document is main guide to write SCORE, so almost things should be in this documents, even it may be duplicated. This document should cover from creating projects with tbears, to deplying contract to main-net.
How to check eventlog recorded.
Example of event handling using both revert() and IconServiceBaseException. It would be better if we explain why we recommend to use revert() rather than IconServiceBaseException.
In InterfaceScore, we need to show the example of other callee Contract's codes.